requirements.txt
Amino.py
futures